Fix: throwing exception error 500 and refactoring the code in the DynamicCatalogStores class
