--- presto-main/src/main/java/io/prestosql/metadata/ExecuteApiRequest.java
@@ -1,9 +1,0 @@
-package io.prestosql.metadata;
-
-import java.util.List;
-
-public interface ExecuteApiRequest
-{
-    public List<DataConnection> executeApiRequest(String dataConnectionEndpoint, String queryParameters);
-}
-
